refactor: move image helpers to service
